refactor(XMLHttpRequest): Use implementation and generated wrapper
