+++ lib/jsdom/browser/Window.js
@@ -13,7 +13,6 @@ const postMessage = require("../living/post-message");
 const DOMException = require("domexception");
 const { btoa, atob } = require("abab");
 const idlUtils = require("../living/generated/utils");
-const createXMLHttpRequest = require("../living/xmlhttprequest");
 const WebSocketImpl = require("../living/websockets/WebSocket-impl").implementation;
 const BarProp = require("../living/generated/BarProp");
 const Document = require("../living/generated/Document");
@@ -436,8 +435,6 @@ function Window(options) {
     }
     return result;
   };
-
-  this.XMLHttpRequest = createXMLHttpRequest(this);
 
   // TODO: necessary for Blob and FileReader due to different-globals weirdness; investigate how to avoid this.
   this.ArrayBuffer = ArrayBuffer;
