+++ lib/jsdom/living/helpers/events.js
@@ -14,7 +14,7 @@ function createAnEvent(e, globalObject, eventInterface = Event, attributes = {})
 function fireAnEvent(e, target, eventInterface, attributes, legacyTargetOverrideFlag) {
   const event = createAnEvent(e, target._globalObject, eventInterface, attributes);
 
-  // tryImplForWrapper() is currently required due to use in Window.js and xmlhttprequest.js
+  // tryImplForWrapper() is currently required due to use in Window.js
   return tryImplForWrapper(target)._dispatch(event, legacyTargetOverrideFlag);
 }
