+++ lib/jsdom/living/interfaces.js
@@ -158,6 +158,7 @@ const generatedInterfaces = [
   require("./generated/FormData"),
   require("./generated/XMLHttpRequestEventTarget"),
   require("./generated/XMLHttpRequestUpload"),
+  require("./generated/XMLHttpRequest"),
   require("./generated/WebSocket"),
 
   require("./generated/NodeIterator"),
