+++ lib/jsdom/living/xhr/xhr-sync-worker.js
@@ -1,12 +1,13 @@
 "use strict";
 /* eslint-disable no-process-exit */
 const util = require("util");
-const { JSDOM } = require("../../..");
+const { JSDOM } = require("../../../api");
+const idlUtils = require("../generated/utils");
 const tough = require("tough-cookie");
-const xhrSymbols = require("./xmlhttprequest-symbols.js");
 
 const dom = new JSDOM();
 const xhr = new dom.window.XMLHttpRequest();
+const xhrImpl = idlUtils.implForWrapper(xhr);
 
 const chunks = [];
 
@@ -26,9 +27,9 @@ process.stdin.on("end", () => {
   }
 
   flag.synchronous = false;
-  xhr[xhrSymbols.flag] = flag;
-  const properties = xhr[xhrSymbols.properties];
-  properties.readyState = xhr.OPENED;
+  xhrImpl.flag = flag;
+  const { properties } = xhrImpl;
+  xhrImpl.readyState = xhr.OPENED;
   try {
     xhr.addEventListener("loadend", () => {
       if (properties.error) {
