+++ lib/jsdom/living/xhr/xhr-utils.js
@@ -6,12 +6,11 @@ const { URL } = require("whatwg-url");
 const parseDataURL = require("data-urls");
 const DOMException = require("domexception");
 
-const ProgressEvent = require("./generated/ProgressEvent");
-const idlUtils = require("./generated/utils");
+const ProgressEvent = require("../generated/ProgressEvent");
+const idlUtils = require("../generated/utils");
 
-const xhrSymbols = require("./xmlhttprequest-symbols");
-const wrapCookieJarForRequest = require("./helpers/wrap-cookie-jar-for-request");
-const { fireAnEvent } = require("./helpers/events");
+const wrapCookieJarForRequest = require("../helpers/wrap-cookie-jar-for-request");
+const { fireAnEvent } = require("../helpers/events");
 
 const headerListSeparatorRegexp = /,[ \t]*/;
 const simpleMethods = new Set(["GET", "HEAD", "POST"]);
@@ -49,7 +48,7 @@ function updateRequestHeader(requestHeaders, header, newValue) {
 }
 
 function dispatchError(xhr) {
-  const errMessage = xhr[xhrSymbols.properties].error;
+  const errMessage = xhr.properties.error;
   requestErrorSteps(xhr, "error", new DOMException(errMessage, "NetworkError"));
 
   if (xhr._ownerDocument) {
@@ -97,8 +96,7 @@ function validCORSPreflightHeaders(xhr, response, flag, properties) {
 }
 
 function requestErrorSteps(xhr, event, exception) {
-  const properties = xhr[xhrSymbols.properties];
-  const flag = xhr[xhrSymbols.flag];
+  const { flag, properties } = xhr;
 
   properties.readyState = xhr.DONE;
   properties.send = false;
@@ -126,18 +124,17 @@ function requestErrorSteps(xhr, event, exception) {
 }
 
 function setResponseToNetworkError(xhr) {
-  const properties = xhr[xhrSymbols.properties];
+  const { properties } = xhr;
   properties.responseCache = properties.responseTextCache = properties.responseXMLCache = null;
   properties.responseHeaders = {};
-  properties.status = 0;
-  properties.statusText = "";
+  xhr.status = 0;
+  xhr.statusText = "";
 }
 
 // return a "request" client object or an event emitter matching the same behaviour for unsupported protocols
 // the callback should be called with a "request" response object or an event emitter matching the same behaviour too
 function createClient(xhr) {
-  const flag = xhr[xhrSymbols.flag];
-  const properties = xhr[xhrSymbols.properties];
+  const { flag, properties } = xhr;
   const urlObj = new URL(flag.uri);
   const uri = urlObj.href;
   const ucMethod = flag.method.toUpperCase();
