--- lib/jsdom/living/xmlhttprequest-symbols.js
@@ -1,4 +1,0 @@
-"use strict";
-
-exports.flag = Symbol("flag");
-exports.properties = Symbol("properties");
